CVE-2016-6317, unsafe query generation risk in AR
There is a vulnerability when Active Record is used in conjunction with
JSON parameter parsing. This vulnerability has been assigned the CVE
identifier CVE-2016-6317.

Due to the way Active Record interprets parameters in combination with
the way that JSON parameters are parsed, it is possible for an attacker
to issue unexpected database queries with "IS NULL" or empty where
clauses.  This issue does *not* let an attacker insert arbitrary values
into an SQL query, however they can cause the query to check for NULL or
eliminate a WHERE clause when most users wouldn't expect it.